refactor(rum-explorer): move utils and some of the data crunching into a separate file
