+++ tools/rum/slicer.js
@@ -1,5 +1,7 @@
 import { fetchPlaceholders } from '../../scripts/lib-franklin.js';
+import { filterBundle, DataChunks } from './cruncher.js';
 import DataLoader from './loader.js';
+import { toHumanReadable, UA_KEY, scoreCWV } from './utils.js';
 
 const mainInnerHTML = `<div class="output">
 <div class="title">
@@ -78,7 +80,6 @@ let chart;
 const BUNDLER_ENDPOINT = 'https://rum.fastly-aem.page/bundles';
 // const BUNDLER_ENDPOINT = 'http://localhost:3000';
 const API_ENDPOINT = BUNDLER_ENDPOINT;
-const UA_KEY = 'userAgent';
 // const API_ENDPOINT = 'https://rum-bundles-2.david8603.workers.dev/rum-bundles';
 // const UA_KEY = 'user_agent';
 
@@ -89,48 +90,7 @@ const canvas = document.getElementById('time-series');
 const timezoneElement = document.getElementById('timezone');
 const lowDataWarning = document.getElementById('low-data-warning');
 
-let dataChunks = [];
-
-/* helpers */
-
-function scoreValue(value, ni, poor) {
-  if (value >= poor) return 'poor';
-  if (value >= ni) return 'ni';
-  return 'good';
-}
-
-function scoreCWV(value, name) {
-  const limits = {
-    lcp: [2500, 4000],
-    cls: [0.1, 0.25],
-    inp: [200, 500],
-    ttfb: [800, 1800],
-  };
-  return scoreValue(value, ...limits[name]);
-}
-
-function toHumanReadable(num) {
-  const dp = 3;
-  let number = num;
-  const thresh = 1000;
-
-  if (Math.abs(num) < thresh) {
-    const precision = (Math.log10(number) < 0) ? 2 : (dp - 1) - Math.floor(Math.log10(number));
-    return `${number.toFixed(precision)}`;
-  }
-
-  const units = ['k', 'm', 'g', 't', 'p'];
-  let u = -1;
-  const r = 10 ** dp;
-
-  do {
-    number /= thresh;
-    u += 1;
-  } while (Math.round(Math.abs(number) * r) / r >= thresh && u < units.length - 1);
-
-  const precision = (dp - 1) - Math.floor(Math.log10(number));
-  return `${number.toFixed(precision)}${units[u]}`;
-}
+const dataChunks = new DataChunks();
 
 function toISOStringWithTimezone(date) {
   // Pad a number to 2 digits
@@ -154,190 +114,6 @@ function setDomain(domain, key) {
   DOMAIN_KEY = key;
   loader.domain = domain;
   loader.domainKey = key;
-}
-
-/* filter, slice and dice bundles */
-
-function filterBundle(bundle, filter, facets, cwv) {
-  let matchedAll = true;
-  const filterMatches = {};
-
-  /* create sub facets */
-  filter.checkpoint.forEach((cp) => {
-    if (!facets[`${cp}.value`]) {
-      facets[`${cp}.value`] = {};
-      cwv[`${cp}.value`] = {};
-    }
-    if (!facets[`${cp}.target`]) {
-      facets[`${cp}.target`] = {};
-      cwv[`${cp}.target`] = {};
-    }
-    if (!facets[`${cp}.source`]) {
-      facets[`${cp}.source`] = {};
-      cwv[`${cp}.source`] = {};
-    }
-  });
-
-  filterMatches.text = true;
-
-  const checkpointEvents = {};
-  const checkpoints = bundle.events.map((e) => {
-    if (!checkpointEvents[e.checkpoint]) checkpointEvents[e.checkpoint] = [];
-    checkpointEvents[e.checkpoint].push(e);
-    return (e.checkpoint);
-  }).filter((cp, index, array) => array.indexOf(cp) === index);
-
-  /* fulltext filter */
-  const fullText = `${bundle.url} ${checkpoints.join()}`;
-  if (!fullText.includes(filter.text)) {
-    matchedAll = false;
-    filterMatches.text = false;
-  }
-
-  /* filter checkpoint */
-  if (matchedAll) {
-    if (filter.checkpoint.length) {
-      if (filter.checkpoint.every((cp) => checkpoints.includes(cp))) {
-        filter.checkpoint.forEach((cp) => {
-          const props = ['source', 'target', 'value'];
-          props.forEach((prop) => {
-            if (filter[`${cp}.${prop}`]) {
-              let anyEventMatchedPropValue = false;
-              const propFilters = filter[`${cp}.${prop}`];
-              checkpointEvents[cp].forEach((cpEvent) => {
-                const propValue = (cp.startsWith('cwv-') && prop === 'value') ? scoreCWV(cpEvent[prop], cp.split('-')[1]) : `${cpEvent[prop]}`;
-                if (propFilters.includes(propValue)) anyEventMatchedPropValue = true;
-              });
-              filterMatches[`${cp}.${prop}`] = anyEventMatchedPropValue;
-              if (!anyEventMatchedPropValue) matchedAll = false;
-            }
-          });
-        });
-        filterMatches.checkpoint = true;
-      } else {
-        matchedAll = false;
-        filterMatches.checkpoint = false;
-      }
-    }
-  }
-
-  /* filter user_agent */
-  if (filter[UA_KEY].length) {
-    if (filter[UA_KEY].includes(bundle[UA_KEY])) {
-      filterMatches[UA_KEY] = true;
-    } else {
-      matchedAll = false;
-      filterMatches[UA_KEY] = false;
-    }
-  }
-
-  /* filter url */
-  if (filter.url.length) {
-    if (filter.url.includes(bundle.url)) {
-      filterMatches.url = true;
-    } else {
-      matchedAll = false;
-      filterMatches.url = false;
-    }
-  }
-
-  const matchedEverythingElse = (facetName) => {
-    let includeInFacet = true;
-    Object.keys(filterMatches).forEach((filterKey) => {
-      if (filterKey !== facetName && !filterMatches[filterKey]) includeInFacet = false;
-    });
-    return includeInFacet;
-  };
-
-  const addToCWV = (facet, option) => {
-    const addMetric = (metric) => {
-      if (!cwv[facet][option]) cwv[facet][option] = {};
-      if (!cwv[facet][option][metric]) cwv[facet][option][metric] = { weight: 0, bundles: [] };
-      const m = cwv[facet][option][metric];
-      m.bundles.push(bundle);
-      m.weight += bundle.weight;
-    };
-    if (bundle.cwvLCP) addMetric('lcp');
-    if (bundle.cwvCLS) addMetric('cls');
-    if (bundle.cwvINP) addMetric('inp');
-    if (bundle.cwvTTFB) addMetric('ttfb');
-  };
-
-  /* facets */
-  if (matchedAll) {
-    checkpoints.forEach((val) => {
-      if (facets.checkpoint[val]) facets.checkpoint[val] += bundle.weight;
-      else facets.checkpoint[val] = bundle.weight;
-      addToCWV('checkpoint', val);
-      if (filter.checkpoint.includes(val)) {
-        const facetOptionsAdded = [];
-        checkpointEvents[val].forEach((e) => {
-          if (e.target) {
-            const facetName = `${val}.target`;
-            const facet = facets[facetName];
-            const option = e.target;
-
-            const facetOptionName = `${facetName}=${option}`;
-            if (!facetOptionsAdded.includes(facetOptionName)) {
-              facetOptionsAdded.push(facetOptionName);
-              if (facet[option]) {
-                facet[option] += bundle.weight;
-              } else {
-                facet[option] = bundle.weight;
-              }
-              addToCWV(facetName, option);
-            }
-          }
-          if (e.source) {
-            const facetName = `${val}.source`;
-            const facet = facets[facetName];
-            const option = e.source;
-
-            const facetOptionName = `${facetName}=${option}`;
-            if (!facetOptionsAdded.includes(facetOptionName)) {
-              facetOptionsAdded.push(facetOptionName);
-              if (facet[option]) {
-                facet[option] += bundle.weight;
-              } else {
-                facet[option] = bundle.weight;
-              }
-              addToCWV(facetName, option);
-            }
-          }
-          if (e.value) {
-            const facetName = `${val}.value`;
-            const facet = facets[facetName];
-            const option = val.startsWith('cwv-') ? scoreCWV(e.value, val.split('-')[1]) : e.value;
-
-            const facetOptionName = `${facetName}=${option}`;
-            if (!facetOptionsAdded.includes(facetOptionName)) {
-              facetOptionsAdded.push(facetOptionName);
-              if (facet[option]) {
-                facet[option] += bundle.weight;
-              } else {
-                facet[option] = bundle.weight;
-              }
-              addToCWV(facetName, option);
-            }
-          }
-        });
-      }
-    });
-  }
-
-  if (matchedEverythingElse('url')) {
-    if (facets.url[bundle.url]) facets.url[bundle.url] += bundle.weight;
-    else facets.url[bundle.url] = bundle.weight;
-    addToCWV('url', bundle.url);
-  }
-
-  if (matchedEverythingElse(UA_KEY)) {
-    if (facets[UA_KEY][bundle[UA_KEY]]) facets[UA_KEY][bundle[UA_KEY]] += bundle.weight;
-    else facets[UA_KEY][bundle[UA_KEY]] = bundle.weight;
-    addToCWV(UA_KEY, bundle[UA_KEY]);
-  }
-
-  return (matchedAll);
 }
 
 /* update UX */
@@ -762,7 +538,6 @@ async function draw() {
   const focus = params.get('focus');
 
   const filterText = params.get('filter') || '';
-  const filtered = [];
   const filter = {
     text: filterText,
     checkpoint,
@@ -790,10 +565,7 @@ async function draw() {
   const startTime = new Date();
   const cwv = structuredClone(facets);
 
-  dataChunks.forEach((chunk) => {
-    filtered.push(...chunk.rumBundles
-      .filter((bundle) => filterBundle(bundle, filter, facets, cwv)));
-  });
+  const filtered = dataChunks.filter((bundle) => filterBundle(bundle, filter, facets, cwv));
 
   if (filtered.length < 1000) {
     lowDataWarning.ariaHidden = 'false';
@@ -873,13 +645,13 @@ async function loadData(scope) {
   const endDate = params.get('endDate') ? `${params.get('endDate')}T00:00:00` : null;
 
   if (scope === 'week') {
-    dataChunks = await loader.fetchLastWeek(endDate);
+    dataChunks.load(await loader.fetchLastWeek(endDate));
   }
   if (scope === 'month') {
-    dataChunks = await loader.fetchPrevious31Days(endDate);
+    dataChunks.load(await loader.fetchPrevious31Days(endDate));
   }
   if (scope === 'year') {
-    dataChunks = await loader.fetchPrevious12Months(endDate);
+    dataChunks.load(await loader.fetchPrevious12Months(endDate));
   }
 
   draw();
