+++ tools/rum/utils.js
@@ -1,0 +1,38 @@
+/* helpers */
+export function scoreValue(value, ni, poor) {
+  if (value >= poor) return 'poor';
+  if (value >= ni) return 'ni';
+  return 'good';
+}
+
+export function scoreCWV(value, name) {
+  const limits = {
+    lcp: [2500, 4000],
+    cls: [0.1, 0.25],
+    inp: [200, 500],
+    ttfb: [800, 1800],
+  };
+  return scoreValue(value, ...limits[name]);
+}
+export const UA_KEY = 'userAgent'; export function toHumanReadable(num) {
+  const dp = 3;
+  let number = num;
+  const thresh = 1000;
+
+  if (Math.abs(num) < thresh) {
+    const precision = (Math.log10(number) < 0) ? 2 : (dp - 1) - Math.floor(Math.log10(number));
+    return `${number.toFixed(precision)}`;
+  }
+
+  const units = ['k', 'm', 'g', 't', 'p'];
+  let u = -1;
+  const r = 10 ** dp;
+
+  do {
+    number /= thresh;
+    u += 1;
+  } while (Math.round(Math.abs(number) * r) / r >= thresh && u < units.length - 1);
+
+  const precision = (dp - 1) - Math.floor(Math.log10(number));
+  return `${number.toFixed(precision)}${units[u]}`;
+}
